Rename java package of the utility.
